Early instantiate MinecraftClientAccessor to avoid NPE.
