+++ common/src/main/java/io/github/reserveword/imblocker/common/accessor/MinecraftClientAccessor.java
@@ -1,11 +1,12 @@
 package io.github.reserveword.imblocker.common.accessor;
 
+import io.github.reserveword.imblocker.common.ReflectionUtil;
 import io.github.reserveword.imblocker.common.gui.Dimension;
 import io.github.reserveword.imblocker.common.gui.Rectangle;
 
 public abstract class MinecraftClientAccessor {
 	
-	public static MinecraftClientAccessor INSTANCE;
+	public static final MinecraftClientAccessor INSTANCE;
 	
 	public abstract void sendSafeCharForFocusTracking(int codePoint);
 	public abstract void execute(Runnable runnable);
@@ -13,5 +14,14 @@ public abstract class MinecraftClientAccessor {
 	public abstract Dimension getContentSize();
 	public abstract Object getCurrentScreen();
 	public abstract int getStringWidth(String text);
-	public abstract void registerClientTickEvent(Runnable tickEvent);
+	
+	static {
+		Class<?> minecraftClientAccessorCls = null;
+		try {
+			minecraftClientAccessorCls = Class.forName("io.github.reserveword.imblocker.MinecraftClientAccessorImpl");
+		} catch (Exception e) {
+			e.printStackTrace();
+		}
+		INSTANCE = (MinecraftClientAccessor) ReflectionUtil.newInstance(minecraftClientAccessorCls, new Class[0]);
+	}
 }
