+++ common/src/main/java/io/github/reserveword/imblocker/common/accessor/ModLoaderAccessor.java
@@ -5,6 +5,7 @@ public interface ModLoaderAccessor {
 	boolean hasMod(String modid);
 	boolean isGameVersionReached(int protocolVersion);
 	Mapping getMapping();
+	void registerClientTickEvent(Runnable tickEvent);
 	
 	public enum Mapping {
 		OFFICIAL, YARN
