+++ forge-legacy/src/main/java/io/github/reserveword/imblocker/ModLoaderAccessorImpl.java
@@ -1,6 +1,10 @@
 package io.github.reserveword.imblocker;
 
 import io.github.reserveword.imblocker.common.accessor.ModLoaderAccessor;
+import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.event.TickEvent;
+import net.minecraftforge.event.TickEvent.Phase;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.loading.FMLLoader;
 
 public class ModLoaderAccessorImpl implements ModLoaderAccessor {
@@ -20,5 +24,17 @@ public class ModLoaderAccessorImpl implements ModLoaderAccessor {
 	@Override
 	public Mapping getMapping() {
 		return Mapping.OFFICIAL;
+	}
+	
+	@Override
+	public void registerClientTickEvent(Runnable tickEvent) {
+		MinecraftForge.EVENT_BUS.register(new Object() {
+			@SubscribeEvent
+			public void onStartTick(TickEvent.ClientTickEvent e) {
+				if(e.phase == Phase.START) {
+					tickEvent.run();
+				}
+			}
+		});
 	}
 }
