+++ forge/src/main/java/io/github/reserveword/imblocker/ModLoaderAccessorImpl.java
@@ -7,6 +7,10 @@ import io.github.reserveword.imblocker.common.IMBlockerCore;
 import io.github.reserveword.imblocker.common.accessor.ModLoaderAccessor;
 import net.minecraft.DetectedVersion;
 import net.minecraft.util.GsonHelper;
+import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.event.TickEvent;
+import net.minecraftforge.event.TickEvent.Phase;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.loading.FMLLoader;
 
 public class ModLoaderAccessorImpl implements ModLoaderAccessor {
@@ -28,6 +32,18 @@ public class ModLoaderAccessorImpl implements ModLoaderAccessor {
 	@Override
 	public Mapping getMapping() {
 		return Mapping.OFFICIAL;
+	}
+	
+	@Override
+	public void registerClientTickEvent(Runnable tickEvent) {
+		MinecraftForge.EVENT_BUS.register(new Object() {
+			@SubscribeEvent
+			public void onStartTick(TickEvent.ClientTickEvent e) {
+				if(e.phase == Phase.START) {
+					tickEvent.run();
+				}
+			}
+		});
 	}
 	
 	static {
