+++ gradle.properties
@@ -11,7 +11,7 @@ minecraft_version_range=1.17+
 # mod
 mod_id=imblocker
 mod_name=IMBlocker
-mod_version=5.4.1
+mod_version=5.4.1.1
 mod_logo=imblocker.png
 mod_description=Blocks Input Methods when you're not using it.\\n在你不打字的时候自动关闭输入法
 mod_url=https://github.com/reserveword/IMBlocker
